split coreplex off into separate package
